Clases de dominio
Ingrediente, IngredienteProducto, Mesa, Comanda (faltan relaciones).

+++ SistemaRestaurante/SistemaRestauranteDominio/src/main/java/sistemarestaurantedominio/dtos/NuevoProductoDTO.java
@@ -25,7 +25,7 @@ public class NuevoProductoDTO {
         return nombre;
     }
 
-    public Float getPrecio() {
+    public double getPrecio() {
         return precio;
     }
